implement memory estimator

+++ server/src/main/java/org/apache/iotdb/db/metadata/mtree/store/disk/memcontrol/IMemManager.java
@@ -44,7 +44,7 @@ public interface IMemManager {
 
   void clear();
 
-  int getPinnedSize();
+  double getPinnedSize();
 
-  int getCachedSize();
+  double getCachedSize();
 }
